backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
httpx==0.28.1
itsdangerous==2.2.0
cryptography==44.0.0
apscheduler==3.10.4
librosa==0.10.2
numpy<2.0
networkx>=3.2
scikit-learn>=1.4
redis[hiredis]>=5.0
python-json-logger>=3.2.1
